add extension methods for non-empty arrays
